requirements.txt
asgiref==3.11.1
Django==4.2.29
djangorestframework==3.16.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
typing_extensions==4.15.0

python-decouple==3.8
honcho==1.1.0
django-filter>=24.0
djangorestframework-simplejwt>=5.5.1
django-admin-confirm>=1.0.1
Pillow>=8.4.0
django-oauth-toolkit>=2.0.0
stripe>=7.0.0
drf-spectacular>=0.27.0
gunicorn>=21.0.0
whitenoise>=6.0.0